pyproject.toml
[project]
name = "llm-wiki-template"
version = "0.0.0"
description = "LLM-maintained markdown wiki template (validation, MkDocs, tests)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mkdocs==1.6.1",
    "mkdocs-material==9.6.7",
    "pymdown-extensions==10.14.3",
    "pytest==8.3.5",
    "PyYAML==6.0.2",
]

[tool.uv]
package = false
